% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ai_in_staff_surveys}
\alias{ai_in_staff_surveys}
\title{\verb{AI in staff surveys} data set}
\format{
A data frame with 3,545 rows and 31 columns:
\describe{
\item{age2}{Participant's age in years.}
\item{gender2}{Participant's gender.}
\item{income2}{Participant's income category.}
\item{location2}{Participant's location (categorical).}
\item{education2}{Participant's level of education.}
\item{ethnicity2}{Participant's ethnicity.}
\item{urban2}{Participant's urbanicity.}
\item{employment2}{Participant's employment status}
\item{aiInt}{Description}
\item{barriers_data}{Description}
\item{barriers_anon}{Description}
\item{barriers_bias}{Description}
\item{barriers_conf}{Description}
\item{barriers_long}{Description}
\item{barriers_other}{Description}
\item{barriers_other_comment}{Description}
\item{reassure_data}{Description}
\item{reassure_human}{Description}
\item{reassure_anon}{Description}
\item{reassure_bias}{Description}
\item{reassure_conf}{Description}
\item{reassure_long}{Description}
\item{reassure_nota}{Description}
\item{reassure_other}{Description}
\item{reassure_other_comment}{Description}
\item{interview_time}{Time spent on questionnaire}
\item{current_page}{Page number at survey completion; completed survey have
\code{completed}}
\item{ip_address}{Hashed IP address of participant}
\item{pid}{Participant identifier; is \code{test} for survey responses via
Predictiv's "Test survey" function}
\item{eligible}{Boolean; whether participant meets eligibility criteria}
\item{attention}{Boolean; whether participant passed attention check}
}
}
\usage{
ai_in_staff_surveys
}
\description{
Data was collected as part of a series of BIT-internal series of mini
experiments. Data set contains raw, unprocessed data.
}
\keyword{datasets}
