% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{c_b_related_policy_hot_topics}
\alias{c_b_related_policy_hot_topics}
\title{\verb{C&B related policy hot topics} data set}
\format{
A data frame with 6,719 rows and 51 columns:
\describe{
\item{age2}{Participant's age in years.}
\item{gender2}{Participant's gender.}
\item{income2}{Participant's income category.}
\item{location2}{Participant's location (categorical).}
\item{education2}{Participant's level of education.}
\item{ethnicity2}{Participant's ethnicity.}
\item{urban2}{Participant's urbanicity.}
\item{employment2}{Participant's employment status}
\item{parent}{Description}
\item{childAge_under1}{Description}
\item{childAge_1}{Description}
\item{childAge_2}{Description}
\item{childAge_3}{Description}
\item{childAge_4}{Description}
\item{childAge_5}{Description}
\item{childAge_6}{Description}
\item{childAge_7}{Description}
\item{childAge_8}{Description}
\item{childAge_9}{Description}
\item{childAge_10}{Description}
\item{childAge_11}{Description}
\item{childAge_12}{Description}
\item{childAge_13}{Description}
\item{childAge_14}{Description}
\item{childAge_15}{Description}
\item{childAge_16}{Description}
\item{childAge_17}{Description}
\item{adultage_declare}{Description}
\item{adultage_doc}{Description}
\item{adultage_image}{Description}
\item{adultage_none}{Description}
\item{adultage_other}{Description}
\item{adultage_other_comment}{Description}
\item{adultVPN}{Description}
\item{childAI_advice}{Description}
\item{childAI_support}{Description}
\item{childAI_search}{Description}
\item{childAI_study}{Description}
\item{childAI_creative}{Description}
\item{childAI_idkuse}{Description}
\item{childAI_idkai}{Description}
\item{childAI_none}{Description}
\item{childAI_other}{Description}
\item{childAI_other_comment}{Description}
\item{childVPN}{Description}
\item{interview_time}{Time spent on questionnaire}
\item{current_page}{Page number at survey completion; completed survey have
\code{completed}}
\item{ip_address}{Hashed IP address of participant}
\item{pid}{Participant identifier; is \code{test} for survey responses via
Predictiv's "Test survey" function}
\item{eligible}{Boolean; whether participant meets eligibility criteria}
\item{attention}{Boolean; whether participant passed attention check}
}
}
\usage{
c_b_related_policy_hot_topics
}
\description{
Data was collected as part of a series of BIT-internal series of mini
experiments. Data set contains raw, unprocessed data.
}
\keyword{datasets}
