% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_from_db}
\alias{import_from_db}
\title{Import data from a database}
\usage{
import_from_db(db = "ducklake", query, credentials)
}
\arguments{
\item{db}{A string with the database from which to fetch results}

\item{query}{String with SQL query}

\item{credentials}{List with database credentials. See the \strong{Credentials}
section for required fields per database.}
}
\value{
tibble with requested data
}
\description{
Currently supported values for \code{db}: \code{"ducklake"}.
If you don't know what to enter as database credentials, please reach out to
the Predictiv team, so they can assign you an account.
}
\section{Credentials}{


\strong{\code{"ducklake"}}\tabular{ll}{
   Field \tab Description \cr
   \code{s3_key_id} \tab S3 access key ID \cr
   \code{s3_secret} \tab S3 secret access key \cr
   \code{s3_endpoint} \tab S3 endpoint \cr
   \code{s3_region} \tab S3 region \cr
   \code{pg_user} \tab Postgres username \cr
   \code{pg_pw} \tab Postgres password \cr
   \code{pg_host} \tab Postgres host \cr
   \code{pg_port} \tab Postgres port \cr
   \code{pg_schema} \tab Postgres schema containing DuckLake metadata \cr
   \code{lake_data_path} \tab S3 path to Parquet data files \cr
   \code{lake_metadata_schema} \tab DuckLake catalog name used in queries \cr
}
}

\examples{
\dontrun{
db <- "ducklake"
query <- "SELECT * FROM lake_catalog.predictiv.results"
credentials <- list(
  s3_key_id= "test",
  s3_secret = "test",
  s3_endpoint = "localhost:4566",
  s3_region = "us-east-1",
  pg_user = "postgres",
  pg_pw = "postgres",
  pg_host = "localhost",
  pg_port = "5434",
  pg_schema = "ducklake",
  lake_data_path = "s3://test/",
  lake_metadata_schema = "lake_catalog"
)
results <- import_from_db(db, query, credentials)
}
}
