% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_from_file}
\alias{import_from_file}
\title{Import data from a local CSV file}
\usage{
import_from_file(path)
}
\arguments{
\item{path}{A string with the local path to the data file. Must include file ending.}
}
\value{
tibble with the imported data
}
\description{
Import data from a local CSV file
}
\examples{
\dontrun{
path <- "./this_is/the_relative_path/to_my/data.csv"
data <- import_from_file(here::here(path))
}
}
