% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parents_beliefs_towards_use_of_ai}
\alias{parents_beliefs_towards_use_of_ai}
\title{\verb{Parents beliefs towards use of AI} data set}
\format{
A data frame with 6,719 rows and 46 columns:
\describe{
\item{age2}{Participant's age in years.}
\item{gender2}{Participant's gender.}
\item{income2}{Participant's income category.}
\item{location2}{Participant's location (categorical).}
\item{education2}{Participant's level of education.}
\item{ethnicity2}{Participant's ethnicity.}
\item{urban2}{Participant's urbanicity.}
\item{employment2}{Participant's employment status}
\item{parent}{Description}
\item{childAge_under1}{Description}
\item{childAge_1}{Description}
\item{childAge_2}{Description}
\item{childAge_3}{Description}
\item{childAge_4}{Description}
\item{childAge_5}{Description}
\item{childAge_6}{Description}
\item{childAge_7}{Description}
\item{childAge_8}{Description}
\item{childAge_9}{Description}
\item{childAge_10}{Description}
\item{childAge_11}{Description}
\item{childAge_12}{Description}
\item{childAge_13}{Description}
\item{childAge_14}{Description}
\item{childAge_15}{Description}
\item{childAge_16}{Description}
\item{childAge_17}{Description}
\item{scaleorder}{Description}
\item{messagingTreatment}{Description}
\item{teachertreatment}{Description}
\item{reliancetreatment}{Description}
\item{aiConcerns_accuracy}{Description}
\item{aiConcerns_equality}{Description}
\item{aiConcerns_emotionalReliance}{Description}
\item{aiConcerns_dataPrivacy}{Description}
\item{govtMessaging_1}{Description}
\item{govtMessaging_2}{Description}
\item{teacherWorkload_1}{Description}
\item{teacherWorkload_2}{Description}
\item{emotionalReliance}{Description}
\item{interview_time}{Time spent on questionnaire}
\item{current_page}{Page number at survey completion; completed survey have
\code{completed}}
\item{ip_address}{Hashed IP address of participant}
\item{pid}{Participant identifier; is \code{test} for survey responses via
Predictiv's "Test survey" function}
\item{eligible}{Boolean; whether participant meets eligibility criteria}
\item{attention}{Boolean; whether participant passed attention check}
}
}
\usage{
parents_beliefs_towards_use_of_ai
}
\description{
Data was collected as part of a series of BIT-internal series of mini
experiments. Data set contains raw, unprocessed data.
}
\keyword{datasets}
