% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reducing_polarisation}
\alias{reducing_polarisation}
\title{\verb{Reducing polarisation} data set}
\format{
A data frame with 6,719 rows and 22 columns:
\describe{
\item{age2}{Participant's age in years.}
\item{gender2}{Participant's gender.}
\item{income2}{Participant's income category.}
\item{location2}{Participant's location (categorical).}
\item{education2}{Participant's level of education.}
\item{ethnicity2}{Participant's ethnicity.}
\item{urban2}{Participant's urbanicity.}
\item{employment2}{Participant's employment status}
\item{politicalLeaning}{Description}
\item{treatmentSpeech}{Description}
\item{budgetAllocation_enforcement}{Description}
\item{budgetAllocation_legalRoutes}{Description}
\item{budgetAllocation_processing}{Description}
\item{budgetAllocation_community}{Description}
\item{affectivePolarisation}{Description}
\item{willingnessOther}{Description}
\item{interview_time}{Time spent on questionnaire}
\item{current_page}{Page number at survey completion; completed survey have
\code{completed}}
\item{ip_address}{Hashed IP address of participant}
\item{pid}{Participant identifier; is \code{test} for survey responses via
Predictiv's "Test survey" function}
\item{eligible}{Boolean; whether participant meets eligibility criteria}
\item{attention}{Boolean; whether participant passed attention check}
}
}
\usage{
reducing_polarisation
}
\description{
Data was collected as part of a series of BIT-internal series of mini
experiments. Data set contains raw, unprocessed data.
}
\keyword{datasets}
