% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trusting_the_machine}
\alias{trusting_the_machine}
\title{\verb{Trusting the machine} data set}
\format{
A data frame with 6,719 rows and 34 columns:
\describe{
\item{age2}{Participant's age in years.}
\item{gender2}{Participant's gender.}
\item{income2}{Participant's income category.}
\item{location2}{Participant's location (categorical).}
\item{education2}{Participant's level of education.}
\item{ethnicity2}{Participant's ethnicity.}
\item{urban2}{Participant's urbanicity.}
\item{employment2}{Participant's employment status}
\item{healthWork}{Description}
\item{gendertreatment}{Description}
\item{aiComfort}{Description}
\item{physicianTrust}{Description}
\item{adherence_startNow}{Description}
\item{adherence_talkTo}{Description}
\item{adherence_lookUpOnline}{Description}
\item{adherence_lookUpChat}{Description}
\item{adherence_rebook}{Description}
\item{adherence_secondOpinion}{Description}
\item{adherence_decline}{Description}
\item{consentBarrier_privacy}{Description}
\item{consentBarrier_accuracy}{Description}
\item{consentBarrier_deskilling}{Description}
\item{consentBarrier_understand}{Description}
\item{consentBarrier_replace}{Description}
\item{consentBarrier_other}{Description}
\item{consentBarrier_other_comment}{Description}
\item{responsibility}{Description}
\item{sentimentOpen}{Description}
\item{interview_time}{Time spent on questionnaire}
\item{current_page}{Page number at survey completion; completed survey have
\code{completed}}
\item{ip_address}{Hashed IP address of participant}
\item{pid}{Participant identifier; is \code{test} for survey responses via
Predictiv's "Test survey" function}
\item{eligible}{Boolean; whether participant meets eligibility criteria}
\item{attention}{Boolean; whether participant passed attention check}
}
}
\usage{
trusting_the_machine
}
\description{
Data was collected as part of a series of BIT-internal series of mini
experiments. Data set contains raw, unprocessed data.
}
\keyword{datasets}
