% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_template.R
\name{use_template}
\alias{use_template}
\title{Copy templates, bundled with the \code{bittoolbox} package}
\usage{
use_template(
  type = "predictiv_cint_cleaning",
  destination = ".",
  filename = "predictiv_cint_cleaning.R",
  overwrite = FALSE
)
}
\arguments{
\item{type}{Character. Name of the template that should be copied.Defaults to
\code{predictiv_cint_cleaning}}

\item{destination}{Character. Path to the directory where the template should
be copied. Defaults to the current working directory (\code{"."}).}

\item{filename}{Character. Name of the copied template file. Defaults to
\code{"predictiv_cint_cleaning.R"}.}

\item{overwrite}{Logical. Whether to overwrite an existing file. Defaults to
FALSE.}
}
\value{
Invisibly returns the path to the copied file.
}
\description{
Copies a template script, bundled with the \code{bittoolbox} package, to a
specific destination. The following templates are currently included:
\itemize{
\item \code{predictiv_cint_cleaning}: cleaning routines for survey data, collected on
Predictiv via Cint
}
}
\details{
The function looks for a template file bundled with the package and copies
it to the specified location. If a file already exists at the destination
and \code{overwrite = FALSE}, the function aborts with an error message.
}
\examples{
\dontrun{
# Copies the default template to the default location
use_template()

# Copies `my_template` to `./my/destination/my_filename.R`, and overwrites
# existing files with the same name
use_template(
  "my_template",
  "./my/destination/",
  "my_filename.R",
  TRUE
)
}
}
