% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_template.R
\name{use_template}
\alias{use_template}
\title{Copy templates, bundled with the \code{bittoolbox} package}
\usage{
use_template(name, target_path = ".", overwrite = FALSE)
}
\arguments{
\item{name}{Character. Name of the template that should be copied.}

\item{target_path}{Character. Target path where the template will be copied to.}

\item{overwrite}{Logical. Whether to overwrite an existing file. Defaults to
FALSE.}
}
\value{
Invisibly returns the path to the target directory.
}
\description{
Copies a template, bundled with the \code{bittoolbox} package, to the current
working directory. The following templates are currently included:
\itemize{
\item \code{predictiv_cint_cleaning}: cleaning routines for survey data, collected on
Predictiv via Cint
\item \code{apac_quarto}: BIT APAC's Quarto template for "producing technical
documents, formatted to resemble other BIT templates for Google docs, and
advice in the BIT brand bible on Norm."
}
}
\details{
The function looks for a template file bundled with the package and copies
it to the specified location. If a file already exists at the target
and \code{overwrite = FALSE}, the function aborts with an error message.
}
\examples{
\dontrun{
use_template(
  "my_template",
  overwrite = TRUE
)
}
}
